refac: Move function from cli.py -> plugins.py
