+++ src/anaconda_cli_base/plugins.py
@@ -3,6 +3,7 @@ import warnings
 from importlib.metadata import EntryPoint
 from importlib.metadata import entry_points
 from sys import version_info
+from typing import Any
 from typing import Dict
 from typing import List
 from typing import Tuple
@@ -10,6 +11,8 @@ from typing import cast
 
 import typer
 from typer.models import DefaultPlaceholder
+
+from anaconda_cli_base.console import console, select_from_list
 
 log = logging.getLogger(__name__)
 
@@ -43,6 +46,60 @@ AUTH_HANDLER_ALIASES = {
 }
 
 
+def _load_auth_handlers(
+    app: typer.Typer,
+    auth_handlers: Dict[str, typer.Typer],
+    auth_handlers_dropdown: List[str],
+) -> None:
+    def validate_at(ctx: typer.Context, _: Any, choice: str) -> str:
+        show_help = ctx.params.get("help", False) is True
+        if show_help:
+            help_str = ctx.get_help()
+            console.print(help_str)
+            raise typer.Exit()
+
+        if choice is None:
+            if len(auth_handlers_dropdown) > 1:
+                choice = select_from_list("choose destination:", auth_handlers_dropdown)
+            else:
+                # If only one is available, we don't need a picker
+                (choice,) = auth_handlers_dropdown
+
+        elif choice not in auth_handlers:
+            print(
+                f"{choice} is not an allowed value for --at. Use one of {auth_handlers_dropdown}"
+            )
+            raise typer.Abort()
+        return choice
+
+    def _action(
+        ctx: typer.Context,
+        at: str = typer.Option(
+            None, callback=validate_at, help=f"Choose from {auth_handlers_dropdown}"
+        ),
+        help: bool = typer.Option(False, "--help"),
+    ) -> None:
+        handler = auth_handlers[at]
+
+        args = ("--help",) if help else ctx.args
+        return handler(args=[ctx.command.name, *args], obj=ctx.obj)
+
+    help_doc = {
+        "login": "Sign into Anaconda services",
+        "logout": "Sign out from Anaconda services",
+        "whoami": "Display account information",
+    }
+
+    for action in "login", "logout", "whoami":
+        decorator = app.command(
+            action,
+            context_settings={"allow_extra_args": True, "ignore_unknown_options": True},
+            rich_help_panel="Authentication",
+            help=help_doc[action],
+        )
+        decorator(_action)
+
+
 def load_registered_subcommands(app: typer.Typer) -> None:
     """Load all subcommands from plugins."""
     subcommand_entry_points = _load_entry_points_for_group(PLUGIN_GROUP_NAME)
@@ -64,8 +121,6 @@ def load_registered_subcommands(app: typer.Typer) -> None:
 
     if auth_handlers:
         auth_handlers_dropdown = sorted(auth_handler_selectors)
-
-        from .cli import _load_auth_handlers
 
         _load_auth_handlers(
             app=app,
